trivial: Move files into modules
