+++ tests/async_wikipedia_page_test.py
@@ -4,7 +4,7 @@ import wikipediaapi
 from tests.mock_data import async_wikipedia_api_request
 from tests.mock_data import user_agent
 from wikipediaapi._enums import Namespace
-from wikipediaapi.async_wikipedia_page import AsyncWikipediaPage
+from wikipediaapi._page.async_wikipedia_page import AsyncWikipediaPage
 
 
 class TestAsyncWikipediaPageInit:
