+++ tests/test_sync_async_symmetry.py
@@ -352,7 +352,7 @@ class TestSyncAsyncPropertySymmetry:
         """Test that all documented properties are available in both classes."""
         # Get all documented properties from ATTRIBUTES_MAPPING
         documented_props = set(
-            wikipediaapi._base_wikipedia_page.BaseWikipediaPage.ATTRIBUTES_MAPPING.keys()
+            wikipediaapi._page._base_wikipedia_page.BaseWikipediaPage.ATTRIBUTES_MAPPING.keys()
         )
 
         # Check sync page has all properties
