+++ wikipediaapi/_image/__init__.py
@@ -1,0 +1,7 @@
+from .async_wikipedia_image import AsyncWikipediaImage
+from .wikipedia_image import WikipediaImage
+
+__all__ = [
+    "WikipediaImage",
+    "AsyncWikipediaImage",
+]
