+++ wikipediaapi/_image/async_wikipedia_image.py
@@ -8,15 +8,15 @@ exposes all data-fetching as awaitables.
 from collections.abc import Coroutine
 from typing import Any
 
-from ._base_wikipedia_page import NOT_CACHED
-from ._base_wikipedia_page import BaseWikipediaPage
-from ._enums import WikiNamespace
-from ._params.imageinfo_params import ImageInfoParams
-from ._types import ImageInfo
-from .wikipedia_page_section import WikipediaPageSection
+from .._enums import WikiNamespace
+from .._page._base_wikipedia_page import NOT_CACHED
+from .._page.wikipedia_page_section import WikipediaPageSection
+from .._params.imageinfo_params import ImageInfoParams
+from .._types import ImageInfo
+from ._base_wikipedia_image import BaseWikipediaImage
 
 
-class AsyncWikipediaImage(BaseWikipediaPage["AsyncWikipediaImage"]):
+class AsyncWikipediaImage(BaseWikipediaImage):
     """Lazy async representation of a Wikipedia/Commons file page.
 
     Mirrors :class:`~wikipediaapi.WikipediaImage` but exposes all
@@ -82,25 +82,6 @@ class AsyncWikipediaImage(BaseWikipediaPage["AsyncWikipediaImage"]):
         if not self._called["imageinfo"]:
             await self._fetch("imageinfo")
         return int(self._attributes.get("pageid", -1)) > 0 or "known" in self._attributes
-
-    def _compute_base_pageid(self) -> int:
-        """Compute deterministic base page ID from image title.
-
-        :return: absolute hash of title for consistent page ID generation
-        """
-        # Use a stable hash function that produces consistent results across runs
-        import hashlib
-
-        return int(hashlib.sha256(self.title.encode("utf-8")).hexdigest(), 16) % (10**18)
-
-    def _get_pageid(self) -> int:
-        """Return page ID based on imageinfo cache.
-
-        :return: positive integer if image exists, negative integer otherwise
-        """
-        exists = int(self._attributes.get("pageid", -1)) > 0 or "known" in self._attributes
-        base = self._compute_base_pageid()
-        return base if exists else -base
 
     @property
     def pageid(self) -> Coroutine[Any, Any, int]:
