+++ wikipediaapi/_page/_base_wikipedia_page.py
@@ -4,9 +4,9 @@ from typing import Any
 from typing import Generic
 from typing import TypeVar
 
-from ._enums import Namespace
-from ._enums import WikiNamespace
-from ._enums import namespace2int
+from .._enums import Namespace
+from .._enums import WikiNamespace
+from .._enums import namespace2int
 from .wikipedia_page_section import WikipediaPageSection
 
 PageT = TypeVar("PageT", bound="BaseWikipediaPage[Any]")
