+++ wikipediaapi/_page/wikipedia_page.py
@@ -8,15 +8,15 @@ for accessing page content, metadata, and related information.
 from typing import Any
 from typing import cast
 
+from .._pages_dict import ImagesDict
+from .._pages_dict import PagesDict
+from .._params.coordinates_params import CoordinatesParams
+from .._params.images_params import ImagesParams
+from .._types import Coordinate
+from .._types import GeoSearchMeta
+from .._types import SearchMeta
 from ._base_wikipedia_page import NOT_CACHED
 from ._base_wikipedia_page import BaseWikipediaPage
-from ._pages_dict import ImagesDict
-from ._pages_dict import PagesDict
-from ._params.coordinates_params import CoordinatesParams
-from ._params.images_params import ImagesParams
-from ._types import Coordinate
-from ._types import GeoSearchMeta
-from ._types import SearchMeta
 from .wikipedia_page_section import WikipediaPageSection
