+++ wikipediaapi/_pages_dict/async_pages_dict.py
@@ -11,21 +11,21 @@ from typing import TYPE_CHECKING
 from typing import Any
 from typing import cast
 
-from .._base_wikipedia_page import BaseWikipediaPage
 from .._enums import CoordinatesProp
 from .._enums import CoordinateType
 from .._enums import Direction
 from .._enums import WikiCoordinatesProp
 from .._enums import WikiCoordinateType
 from .._enums import WikiDirection
+from .._page._base_wikipedia_page import BaseWikipediaPage
 from .base_pages_dict import _AsyncBatchWiki
 from .pages_dict import PagesDict
 
 if TYPE_CHECKING:
+    from .._page.async_wikipedia_page import AsyncWikipediaPage
     from .._resources import BaseWikipediaResource
     from .._types import Coordinate
     from .._types import GeoPoint
-    from ..async_wikipedia_page import AsyncWikipediaPage
 
 
 class AsyncPagesDict(dict[str, BaseWikipediaPage[Any]]):
