+++ wikipediaapi/_pages_dict/images_dict.py
@@ -12,14 +12,14 @@ from typing import TYPE_CHECKING
 from typing import Any
 from typing import cast
 
-from .._base_wikipedia_page import BaseWikipediaPage
+from .._page._base_wikipedia_page import BaseWikipediaPage
 from .._params.imageinfo_params import _DEFAULT_PROP
 from .base_pages_dict import _SyncImageWiki
 
 if TYPE_CHECKING:
+    from .._image.wikipedia_image import WikipediaImage
     from .._resources import BaseWikipediaResource
     from .._types import ImageInfo
-    from ..wikipedia_image import WikipediaImage
 
 
 class ImagesDict(dict[str, BaseWikipediaPage[Any]]):
