+++ wikipediaapi/_wikipedia/__init__.py
@@ -1,0 +1,7 @@
+from .async_wikipedia import AsyncWikipedia
+from .wikipedia import Wikipedia
+
+__all__ = [
+    "Wikipedia",
+    "AsyncWikipedia",
+]
